clean up pipeline utils a bit
