+++ tests/load/duckdb/test_duckdb_client.py
@@ -14,7 +14,8 @@ from dlt.destinations.impl.duckdb.configuration import (
 )
 from dlt.destinations import duckdb
 
-from tests.load.pipeline.utils import drop_pipeline, assert_table
+from tests.load.pipeline.utils import drop_pipeline
+from tests.pipeline.utils import assert_table
 from tests.utils import patch_home_dir, autouse_test_storage, preserve_environ, TEST_STORAGE_ROOT
 
 # mark all tests as essential, do not remove
