+++ tests/load/pipeline/test_duckdb.py
@@ -6,11 +6,10 @@ from dlt.destinations.exceptions import DatabaseTerminalException
 from dlt.pipeline.exceptions import PipelineStepFailed
 
 from tests.cases import TABLE_UPDATE_ALL_INT_PRECISIONS, TABLE_UPDATE_ALL_TIMESTAMP_PRECISIONS
-from tests.pipeline.utils import airtable_emojis
+from tests.pipeline.utils import airtable_emojis, load_table_counts
 from tests.load.pipeline.utils import (
     destinations_configs,
     DestinationTestConfiguration,
-    load_table_counts,
 )
