+++ tests/load/pipeline/test_filesystem_pipeline.py
@@ -22,8 +22,9 @@ from dlt.destinations.path_utils import create_path
 from tests.load.pipeline.utils import (
     destinations_configs,
     DestinationTestConfiguration,
-    load_table_counts,
 )
+
+from tests.pipeline.utils import load_table_counts
 
 
 skip_if_not_active("filesystem")
