+++ tests/load/pipeline/test_merge_disposition.py
@@ -18,8 +18,7 @@ from dlt.extract import DltResource
 from dlt.sources.helpers.transform import skip_first, take_first
 from dlt.pipeline.exceptions import PipelineStepFailed
 
-from tests.pipeline.utils import assert_load_info
-from tests.load.pipeline.utils import load_table_counts, select_data
+from tests.pipeline.utils import assert_load_info, load_table_counts, select_data
 from tests.load.pipeline.utils import destinations_configs, DestinationTestConfiguration
 
 # uncomment add motherduck tests
