+++ tests/load/pipeline/test_write_disposition_changes.py
@@ -4,8 +4,10 @@ from typing import Any
 from tests.load.pipeline.utils import (
     destinations_configs,
     DestinationTestConfiguration,
-    assert_data_table_counts,
 )
+
+from tests.pipeline.utils import assert_data_table_counts
+
 from tests.pipeline.utils import assert_load_info
 from dlt.pipeline.exceptions import PipelineStepFailed
