+++ tests/pipeline/test_schema_contracts.py
@@ -10,7 +10,7 @@ from dlt.extract import DltResource
 from dlt.pipeline.pipeline import Pipeline
 from dlt.pipeline.exceptions import PipelineStepFailed
 
-from tests.load.pipeline.utils import load_table_counts
+from tests.pipeline.utils import load_table_counts
 from tests.utils import (
     TestDataItemFormat,
     skip_if_not_active,
